Revert changes: benchmark shows regression, not improvement
The EgorBot benchmark results show the changes caused a 1.88x
regression on x64 (AMD EPYC), not the expected improvement.
Main: ~0.29ns, PR: ~0.55ns per call for non-ASCII chars.
On ARM64 results were neutral (ratio ~1.00).

Reverting both changes:
- Char.cs: restore call to CharUnicodeInfo.GetUnicodeCategory((int)c)
- CharUnicodeInfo.cs: remove AggressiveInlining attribute

+++ src/libraries/System.Private.CoreLib/src/System/Char.cs
@@ -954,7 +954,7 @@ namespace System
             {
                 return GetLatin1UnicodeCategory(c);
             }
-            return CharUnicodeInfo.GetUnicodeCategory(c);
+            return CharUnicodeInfo.GetUnicodeCategory((int)c);
         }
 
         public static UnicodeCategory GetUnicodeCategory(string s, int index)
